example-projects/flink-statefun-ridesharing/geo-cell-func/requirements.txt
apache-flink-statefun ~= 3.0.0
fastapi ~= 0.63.0
uvicorn[standard] ~= 0.13.4
requests ~= 2.25.1

# for test
pytest ~= 6.2.1
pytest-asyncio ~= 0.14.0
httpx ~= 0.16.1
asgi_lifespan ~= 1.0.1